Move some` FAB related tests to provider
There are still some tests that are really FAB provider Auth Manager
tests but they were left in the core (which is kind-of expected
remnant of the past. We will eventually find and move all of them but
for now let's just move some of them we are aware of.
