+++ tests/providers/fab/auth_manager/test_security.py
@@ -19,6 +19,7 @@ from __future__ import annotations
 
 import contextlib
 import datetime
+import json
 import logging
 import os
 from unittest import mock
@@ -1139,3 +1140,66 @@ def test_custom_access_denied_message():
     ):
         initialize_config()
         assert get_access_denied_message() == "My custom access denied message"
+
+
+@pytest.mark.db_test
+class TestHasAccessDagDecorator:
+    @pytest.mark.parametrize(
+        "dag_id_args, dag_id_kwargs, dag_id_form, dag_id_json, fail",
+        [
+            ("a", None, None, None, False),
+            (None, "b", None, None, False),
+            (None, None, "c", None, False),
+            (None, None, None, "d", False),
+            ("a", "a", None, None, False),
+            ("a", "a", "a", None, False),
+            ("a", "a", "a", "a", False),
+            (None, "a", "a", "a", False),
+            (None, None, "a", "a", False),
+            ("a", None, None, "a", False),
+            ("a", None, "a", None, False),
+            ("a", None, "c", None, True),
+            (None, "b", "c", None, True),
+            (None, None, "c", "d", True),
+            ("a", "b", "c", "d", True),
+        ],
+    )
+    def test_dag_id_consistency(
+        self,
+        app,
+        dag_id_args: str | None,
+        dag_id_kwargs: str | None,
+        dag_id_form: str | None,
+        dag_id_json: str | None,
+        fail: bool,
+    ):
+        with app.test_request_context() as mock_context:
+            from airflow.www.auth import has_access_dag
+
+            mock_context.request.args = {"dag_id": dag_id_args} if dag_id_args else {}
+            kwargs = {"dag_id": dag_id_kwargs} if dag_id_kwargs else {}
+            mock_context.request.form = {"dag_id": dag_id_form} if dag_id_form else {}
+            if dag_id_json:
+                mock_context.request._cached_data = json.dumps({"dag_id": dag_id_json})
+                mock_context.request._parsed_content_type = ["application/json"]
+
+            with create_user_scope(
+                app,
+                username="test-user",
+                role_name="limited-role",
+                permissions=[(permissions.ACTION_CAN_READ, permissions.RESOURCE_DAG)],
+            ) as user:
+                with patch(
+                    "airflow.providers.fab.auth_manager.fab_auth_manager.FabAuthManager.get_user"
+                ) as mock_get_user:
+                    mock_get_user.return_value = user
+
+                    @has_access_dag("GET")
+                    def test_func(**kwargs):
+                        return True
+
+                    result = test_func(**kwargs)
+                    if fail:
+                        assert result[1] == 403
+                    else:
+                        assert result is True
